Type declarations on global variables are not yet supported in version 1.7

+++ src/Volume.jl
@@ -19,14 +19,14 @@ const Shape{T} = Union{NoShape{T},
 
 #---Volume-------------------------------------------------------------------------
 struct VolumeP{T<:AbstractFloat,PV}
-    id::UInt64
+    id::Int64
     label::String
     shape::Shape{T}                     # Reference to the actual shape
     material::Material{T}               # Reference to material
     daughters::Vector{PV}
 end
 
-_last_volume_id::UInt64 = 1
+_last_volume_id = 1
 
 function VolumeP{T, PV}(label::String, shape::Shape{T}, material::Material{T}, daughters::Vector{PV}) where {T<:AbstractFloat, PV}
     global _last_volume_id
